Create Event entity
Event is a JPA entity.

There is no `Table` annotation, so it is assumed that this entity
is attached to a table of the same name.

`Id` annotation on top of id field ensures that it will be recognized as
object's
id. @GeneratedValue ensures it is generated automatically.

startAt, endAt, and description have no annotations at all, therefore, must
be matched database columns of the same names.